Split byref tests into own file
